# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  mydb.StructAdmin:
    properties:
      avatar:
        description: 用户头像
        type: string
      createTime:
        type: integer
      email:
        type: string
      id:
        type: integer
      lastLoginTime:
        type: integer
      password:
        type: string
      phoneNumber:
        type: string
      role:
        type: string
      salt:
        type: string
      status:
        type: integer
      updateTime:
        type: integer
      username:
        type: string
    type: object
...
  upload.ImgReturnData:
    properties:
      hash:
        type: string
      img_path:
        type: string
    type: object
  util.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: nav.fandoc.org
info:
  contact: {}
  description: This is a sample server for a nav web site.
  title: Nav Web Site API
  version: "1.0"
paths:
  /admin/delete/{id}:
    delete:
      description: 根据用户ID删除用户
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 删除用户
      tags:
      - admin
  /admin/detail/{id}:
    get:
      description: 根据用户ID获取用户的详细信息
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/mydb.StructAdmin'
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 获取用户详情
      tags:
      - admin
  /admin/editProfile/{id}:
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: 根据用户ID编辑用户的资料
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 邮箱地址
        in: formData
        name: email
        type: string
      - description: 手机号
        in: formData
        name: phone_number
        type: string
      - description: 用户头像
        in: formData
        name: avatar
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 编辑用户资料
      tags:
      - admin
  /admin/list:
    get:
      description: 获取所有用户的列表
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  items:
                    allOf:
                    - $ref: '#/definitions/mydb.StructAdmin'
                    - properties:
                        password:
                          type: string
                        salt:
                          type: string
                      type: object
                  type: array
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 获取用户列表
      tags:
      - admin
...
  /admin/updatePassword/{id}:
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: 根据用户ID修改用户的密码
      parameters:
      - description: 认证Token
        in: header
        name: LoginToken
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 新密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 修改用户密码
      tags:
      - admin
...
  /upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片文件
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/upload.ImgReturnData'
                message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/util.APIResponse'
            - properties:
                code:
                  type: integer
                data:
                  type: object
                message:
                  type: string
              type: object
      summary: 上传图片
      tags:
      - upload
swagger: "2.0"